Moved mason installer function to lsp utils
